refactor: Add routes.report. Move branding components to branding.ts.
Splitting out branding.ts avoids potential issues with importing components
from e.g. app/robots.ts, app/sitemap.ts, i18n/request.ts or middleware.ts

Upgrade ids-drr-branding-types:
- Use new aboutUsEnabled configuration
- Move reportsEnabled to features.reports
- Add sideEffects to branding package
