+++ app/[locale]/about-us/page.tsx
@@ -1,6 +1,6 @@
 import { notFound } from 'next/navigation';
 
-import { AboutPage } from '@/config/site';
+import { AboutPage } from '@/config/branding';
 
 export default function Page() {
   if (!AboutPage) notFound();
