+++ branding-stub/package.json
@@ -5,6 +5,7 @@
   "main": "src/index.ts",
   "types": "src/index.ts",
   "private": true,
+  "sideEffects": ["**/*.css", "**/*.scss"],
   "dependencies": {
     "ids-drr-branding-types": "github:open-contracting/ids-drr-branding-types"
   }
