+++ config/site.ts
@@ -1,6 +1,5 @@
-import * as branding from 'ids-drr-branding';
+import { config } from 'ids-drr-branding';
 import type {
-  Exports,
   Language,
   Resource,
   StaticImageAsset,
@@ -9,24 +8,11 @@ import type {
 
 import { routes } from '@/lib/routes';
 
-// Catch missing or wrong-typed exports at compile-time, not runtime.
-const validated: Exports = branding;
-
-const { config } = validated;
-
-// Re-export branding-provided components and values so the rest of the app
-// imports all deployment-specific things from one place.
-export const About = validated.About;
-export const AboutPage = validated.AboutPage;
-export const Credits = validated.Credits;
-export const DataStories = validated.DataStories;
-export const Footer = validated.Footer;
-export const Partners = validated.Partners;
-export const PartnerLogos = validated.PartnerLogos;
 // Arrays
 export const states: State[] = config.states ?? [];
 export const resources: Resource[] = config.resources ?? [];
 export const languages: Language[] = config.languages ?? [];
+
 // Images
 export const logo: StaticImageAsset | undefined = config.logo;
 export const heroForeground: StaticImageAsset | undefined = config.heroForeground;
@@ -34,11 +20,10 @@ export const heroBackground: string = config.heroBackground ?? '';
 export const favicon: string = config.favicon ?? '';
 export const appleIcon: string = config.appleIcon ?? '';
 export const openGraphImage: string = config.openGraphImage ?? '';
+
 // Links
 export const userGuideLink: string = config.userGuideLink ?? '';
 export const docsLink: string = config.docsLink ?? '';
-// Feature flags
-export const reportsEnabled: boolean = config.reportsEnabled ?? false;
 
 // English is always loaded internally as the missing-key fallback,
 // but a deployment can omit it from `locales` to disable /en/ URLs.
@@ -48,15 +33,16 @@ export const defaultLocale: string = config.defaultLocale ?? locales[0];
 export const messages: Record<string, Record<string, unknown>> =
   config.messages ?? {};
 
-// Feature flags.
-const backendAvailable = Boolean(process.env.NEXT_PUBLIC_BACKEND_URL);
+// Feature flags
+const dataSpaceEnabled = Boolean(process.env.NEXT_PUBLIC_BACKEND_URL);
 export const features = {
-  chart: backendAvailable,
-  datasets: backendAvailable,
-  aboutUs: Boolean(AboutPage),
+  chart: dataSpaceEnabled,
+  datasets: dataSpaceEnabled,
+  aboutUs: config.aboutUsEnabled ?? false,
+  reports: config.reportsEnabled ?? false,
 };
 
-// Navigation.
+// Navigation
 const defaultState = states.find((s) => s.status === 'active');
 
 export const siteUrl =
